refactor: rename directories, update imports
